Refactor MediaSigner facade module layout
