+++ pkgs/plugins/media_signer/MediaSigner/_signer.py
