+++ pkgs/plugins/media_signer/MediaSigner/cli.py
@@ -10,7 +10,7 @@ import sys
 from pathlib import Path
 from typing import Any, Mapping, MutableMapping, Sequence
 
-from .Signer import MediaSigner
+from . import MediaSigner
 
 BytesLike = bytes | bytearray
